+++ src/client/components/ac_change_pwd/ac_change_pwd.js
@@ -17,8 +17,47 @@ Template.ac_change_pwd.onCreated( function(){
 
     self.AC = {
         component: new ReactiveVar( null ),
-        passwordOk: new ReactiveVar( true ),
-        twiceOk: new ReactiveVar( true )
+        passwordOk: new ReactiveVar( false ),
+        passwordVal: null,
+        twiceOk: new ReactiveVar( false ),
+        twiceVal: null,
+        checksOk: new ReactiveVar( false ),
+
+        // not only each field must be individually ok
+        //  but we also make sure the old and new passwords are different
+        checks( event, data ){
+            switch( event.type ){
+                case 'ac-password-data':
+                    self.AC.checkPassword( data );
+                    self.AC.checkPanel();
+                    break;
+                case 'ac-twice-data':
+                    self.AC.checkTwice( data );
+                    self.AC.checkPanel();
+                    break;
+            }
+        },
+        // only relevant if individual checks are ok
+        checkPanel(){
+            let isOk = self.AC.passwordOk.get() && self.AC.twiceOk.get();
+            if( isOk ){
+                isOk &&= self.AC.passwordVal && self.AC.twiceVal && self.AC.passwordVal !== self.AC.twiceVal;
+                if( self.AC.passwordVal && self.AC.passwordVal.length > 0 && self.AC.passwordVal === self.AC.twiceVal ){
+                    AccountsUI.Display.errorMsg( pwixI18n.label( I18N, 'change_pwd.pwds_are_equal' ));
+                }
+            }
+            self.AC.checksOk.set( isOk );
+        },
+        checkPassword( data ){
+            //console.debug( data );
+            self.AC.passwordVal = data.password;
+            self.AC.passwordOk.set( data.ok );
+        },
+        checkTwice( data ){
+            //console.debug( data );
+            self.AC.twiceVal = data.password;
+            self.AC.twiceOk.set( data.ok );
+        }
     };
 
     // setup the acUserLogin acManager component
@@ -33,9 +72,15 @@ Template.ac_change_pwd.onCreated( function(){
 Template.ac_change_pwd.onRendered( function(){
     const self = this;
 
+    const $acContent = self.$( '.ac-change-pwd' ).closest( '.ac-content' );
+
     self.autorun(() => {
-        const btn = self.$( '.ac-change-pwd' ).closest( '.ac-content' ).find( '.ac-submit' );
-        btn.prop( 'disabled', !self.AC.passwordOk.get() || !self.AC.twiceOk.get());
+        $acContent.attr( 'data-ac-requester', Template.currentData().managerId );
+    });
+
+    self.autorun(() => {
+        //console.debug( self );
+        $acContent.find( '.ac-submit' ).prop( 'disabled', !self.AC.checksOk.get());
     });
 });
 
@@ -85,14 +130,18 @@ Template.ac_change_pwd.events({
     //  NB: happens that data arrives undefined :( see #24
     'ac-password-data .ac-change-pwd'( event, instance, data ){
         //console.log( 'ac-password-data', data );
-        instance.AC.passwordOk.set( data ? data.ok : false );
         AccountsUI.Display.errorMsg( '' );
+        if( data ){
+            instance.AC.checks( event, data );
+        }
     },
 
     // message sent by the twice passwords component
     'ac-twice-data .ac-change-pwd'( event, instance, data ){
         //console.log( 'ac-twice-data', data );
-        instance.AC.twiceOk.set( data ? data.ok : false );
         AccountsUI.Display.errorMsg( '' );
+        if( data ){
+            instance.AC.checks( event, data );
+        }
     }
 });