+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Upgrade pwix:options requirement to 2.1.0
     - Deep code refactoring and configuration parameters evolution (bumping candidate version number)
     - Remove AccountsUI.preferredLabel() method, moved to pwix:accounts-tools
+    - Make sure old and new passwords are different when changing a password
 
 ### 1.3.0